+++ shibuya-core/test/Main.hs
@@ -12,6 +12,7 @@ import Shibuya.Core.TypesSpec qualified
 import Shibuya.PolicySpec qualified
 import Shibuya.Runner.BatchProcessorSpec qualified
 import Shibuya.Runner.BatcherSpec qualified
+import Shibuya.Runner.PartitionOrderingSpec qualified
 import Shibuya.Runner.SupervisedSpec qualified
 import Shibuya.RunnerSpec qualified
 import Shibuya.Telemetry.EffectSpec qualified
@@ -32,6 +33,7 @@ main = hspec $ do
   describe "Shibuya.Runner" Shibuya.RunnerSpec.spec
   Shibuya.Runner.BatcherSpec.spec
   Shibuya.Runner.BatchProcessorSpec.spec
+  Shibuya.Runner.PartitionOrderingSpec.spec
   Shibuya.Runner.SupervisedSpec.spec
   Shibuya.Telemetry.EffectSpec.spec
   Shibuya.Telemetry.PropagationSpec.spec